Separate expressions to multiple classes
