+++ enterprise/cypher/compiled-expressions/src/main/java/org/neo4j/cypher/internal/runtime/compiled/expressions/CypherMath.java
@@ -65,44 +65,14 @@ import static org.neo4j.values.storable.Values.doubleValue;
 import static org.neo4j.values.storable.Values.stringValue;
 
 /**
- * This class contains static helper methods used by the compiled expressions
+ * This class contains static helper math methods used by the compiled expressions
  */
-public final class ExpressionMethods
+@SuppressWarnings( "unused" )
+public final class CypherMath
 {
-    private static final BooleanMapper BOOLEAN_MAPPER = new BooleanMapper();
-
-    private ExpressionMethods()
+    private CypherMath()
     {
         throw new UnsupportedOperationException( "Do not instantiate" );
-    }
-
-    public static DoubleValue sin( AnyValue in )
-    {
-        if ( in instanceof NumberValue )
-        {
-            return doubleValue( Math.sin( ((NumberValue) in).doubleValue() ) );
-        }
-        else
-        {
-            throw new CypherTypeException( "sin requires numbers", null );
-        }
-    }
-
-    public static DoubleValue round( AnyValue in )
-    {
-        if ( in instanceof NumberValue )
-        {
-            return doubleValue( Math.round( ((NumberValue) in).doubleValue() ) );
-        }
-        else
-        {
-            throw new CypherTypeException( "round requires numbers", null );
-        }
-    }
-
-    public static DoubleValue rand()
-    {
-        return doubleValue( ThreadLocalRandom.current().nextDouble() );
     }
 
     //TODO this is horrible spaghetti code, we should push most of this down to AnyValue
@@ -249,175 +219,5 @@ public final class ExpressionMethods
         }
         throw new CypherTypeException(
                 String.format( "Don't know how to subtract `%s` and `%s`", lhs, rhs ), null );
-    }
-
-    //data access
-    public static Value nodeProperty( Transaction tx, long node, int property )
-    {
-        CursorFactory cursors = tx.cursors();
-        try ( NodeCursor nodes = cursors.allocateNodeCursor();
-              PropertyCursor properties = cursors.allocatePropertyCursor() )
-        {
-            tx.dataRead().singleNode( node, nodes );
-            if ( nodes.next() )
-            {
-                nodes.properties( properties );
-                return property( properties, property );
-            }
-            return Values.NO_VALUE;
-        }
-    }
-
-    public static Value relationshipProperty( Transaction tx, long relationship, int property )
-    {
-        CursorFactory cursors = tx.cursors();
-        try ( RelationshipScanCursor relationships = cursors.allocateRelationshipScanCursor();
-              PropertyCursor properties = cursors.allocatePropertyCursor() )
-        {
-            tx.dataRead().singleRelationship( relationship, relationships );
-            if ( relationships.next() )
-            {
-                relationships.properties( properties );
-                return property( properties, property );
-            }
-            return Values.NO_VALUE;
-        }
-    }
-
-    //boolean operations
-
-    public static Value or( AnyValue... args )
-    {
-        for ( AnyValue arg : args )
-        {
-            if ( arg == NO_VALUE )
-            {
-                return NO_VALUE;
-            }
-
-            if ( arg.map( BOOLEAN_MAPPER ) )
-            {
-                return Values.TRUE;
-            }
-        }
-        return Values.FALSE;
-    }
-
-    private static final class BooleanMapper implements ValueMapper<Boolean>
-    {
-
-        @Override
-        public Boolean mapPath( PathValue value )
-        {
-            return value.size() > 0;
-        }
-
-        @Override
-        public Boolean mapNode( VirtualNodeValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-
-        @Override
-        public Boolean mapRelationship( VirtualRelationshipValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-
-        @Override
-        public Boolean mapMap( MapValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-
-        @Override
-        public Boolean mapNoValue()
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + NO_VALUE, null);
-        }
-
-        @Override
-        public Boolean mapSequence( SequenceValue value )
-        {
-            return value.length() > 0;
-        }
-
-        @Override
-        public Boolean mapText( TextValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-
-        @Override
-        public Boolean mapBoolean( BooleanValue value )
-        {
-            return value.booleanValue();
-        }
-
-        @Override
-        public Boolean mapNumber( NumberValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapDateTime( DateTimeValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapLocalDateTime( LocalDateTimeValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapDate( DateValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-
-        @Override
-        public Boolean mapTime( TimeValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapLocalTime( LocalTimeValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapDuration( DurationValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-
-        }
-
-        @Override
-        public Boolean mapPoint( PointValue value )
-        {
-            throw new CypherTypeException( "Don't know how to treat that as a boolean: " + value, null);
-        }
-    }
-
-    private static Value property( PropertyCursor properties, int property )
-    {
-        while ( properties.next() )
-        {
-            if ( properties.propertyKey() == property )
-            {
-                return properties.propertyValue();
-            }
-        }
-        return Values.NO_VALUE;
     }
 }
